cdk/lib/termin-radar-check/termin-radar-load-service-page-lambda-function/lambda/requirements.txt
aws_lambda_powertools==3.25.0
aws-xray-sdk==2.15.0
boto3==1.42.61
pyhumps==3.8.0
